Restructure endpoints into separate directories
