+++ Vibify/API/Dalle/DalleResponse.swift
