+++ Vibify/API/GPT/GPTResponse.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct GPTResponse: Codable {
+    let choices: [Choice]
+    
+    struct Choice: Codable {
+        let text: String
+    }
+}
